[add, test, fix] Not forcing push to generated
Fetching both the updated branch (aka changes) and generated, then generate the files in the new changes and remove the `.git/` dir. In the `generated/` dir remove everything except the `.git/` dir and move the changes and generated files there. Then push the new changes, this way there is no need to overwrite the history and the changes between commits can be diffed.

Note that also the `.gitignore` on the `changes` dir is removed, to ensure all the generated files are added.