[refactor] Moving literate config to a "normal" config, user
After 4 months using literate config (this was the commit were it all
started 60d4f5b2fac2eebc27e5dd28096c36e885908605) I've been noticed I
didn't really like the setup, it start to get really messy to deal
with, and whenever I have to add new config I have to do the whole
process of making a new section, adding the begin and end tags of the
src block, the names of each block, putting the name where it should
be tangled, etc.

So I will try moving the config to this "normal config" format and see
if I like it more.